small change and one problem
added the email and password through the environment variable. Added a thank you page once the form is submitted. The problem is that when clicking submit, the URL can't be found, the route is defined in views.py and its also shown correctly on the support.html file, but unsure what the problem is, will be fixed soon.