minor update to support/views file
switched the destination of From/To for email to send to the help desk and not the user's email. Also added the text for the user to either use the user or email. This might change in the future, and the support page's email function still doesnt work. More progress later...