updated support code in views file
Added error exception handling for the support page, the function still doesnt work but instead of going to the error screen after attempting to submit form, it will redirect to the support page and flash a message mentioned that a error has occured sending the form. Also included the error details into the terminal instead of the error screen.